Remove a test for checking if dialogue opens inside the game engine.
There isn't actually an assert that this happens. This would require visual regression testing; probably an ok idea to have a little bit of editor/engine integration testing, but mostly this sort of testing should be on the engine side.

Anyhow, this test is flaky; passes in my local, fails in CI. I tried adding a wait between the steps - that made it fail in my local.

These difficulties aside, I think the test is not doing what it's supposed to. Removing for now.